+++ mangadex_downloader/cli/command.py
@@ -43,7 +43,8 @@ from ..iterator import (
 )
 from ..forums import (
     get_thread_title_owner_and_post_owner,
-    get_post_id_forum_thread
+    get_post_id_forum_thread,
+    validate_forum_thread_url
 )
 from .. import __repository__
 from ..utils import input_handle, validate_url, get_cover_art_url
@@ -548,6 +549,7 @@ class ForumThreadCommand(MangaDexCommand):
     def __init__(self, parser, args, input_text):
         iterator = ForumThreadMangaDexURLIterator(input_text, True)
 
+        input_text = validate_forum_thread_url(input_text).url
         post_id = get_post_id_forum_thread(input_text)
         result = get_thread_title_owner_and_post_owner(thread_url=input_text, post_id=post_id)
         thread_title, thread_owner, post_owner = result